One-hot encoding with Rust transform
This patch extends sklearn's OHE with a subclass,
that dispatches the CSR transform to the Rust
backend. stratum reuses sklearn's fit. For transform,
we first recode using Pandas, send the recoded features
to Rust, which allocates and fill a CSR output. If
sparse_output is false, we use a multi-threaded Rust
implementation for densification. Overall stratum is
3x faster.